ansible/requirements.txt
ansible-compat==4.1.11
ansible-core==2.16.6
ansible-lint==24.2.2
attrs==23.2.0
black==24.4.0
bracex==2.4
cffi==1.16.0
click==8.1.7
click-help-colors==0.9.4
cryptography==42.0.5
enrich==1.2.7
filelock==3.13.4
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
molecule==24.2.1
molecule-plugins==23.5.3
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.5.0
pycparser==2.22
Pygments==2.17.2
PyYAML==6.0.1
referencing==0.34.0
resolvelib==1.0.1
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
subprocess-tee==0.4.1
wcmatch==8.5.1
yamllint==1.35.1
iniconfig=2.0.0
pytest=8.1.1
pytest-testinfra=10.1.0
